requirements.txt
rpi_ws281x==5.3.0
flask==3.0.2
python-dotenv==1.0.1 